requirements.txt
quandl
flask==0.11
pandas_datareader